move exceptions in  to exceptions.py
