+++ bust/exceptions.py
@@ -8,7 +8,7 @@ def error(msg):
 
 
 class FormatError(Exception):
-    """! @brief Raised when input JSON is incorrect
+    """! @brief Raised when input JSON is incorrect.
 
 
     """
@@ -16,7 +16,7 @@ class FormatError(Exception):
 
 class InvalidAddress(RuntimeError):
     """! @brief Raised when a specified address is already assigned to
-    other register
+    other register.
 
     """
 
@@ -28,7 +28,7 @@ class InvalidAddress(RuntimeError):
 
 class InvalidRegister(RuntimeError):
     """! @brief Raised when trying to parse a register with lacking or misspelt
-    json keys
+    json keys.
     """
 
     def __init__(self, reg):
@@ -37,7 +37,7 @@ class InvalidRegister(RuntimeError):
         super().__init__(error(msg))
 
 class InvalidBusType(RuntimeError):
-    """ @brief Raised when trying to parse an unspecified or unsupported bus type
+    """ @brief Raised when trying to parse an unspecified or unsupported bus type.
 
     """
     def __init__(self, bus_type):
@@ -46,7 +46,7 @@ class InvalidBusType(RuntimeError):
 
 
 class InvalidResetMode(RuntimeError):
-    """Documentation for InvalidResetMode
+    """Documentation for InvalidResetMode.
 
     """
     def __init__(self, reset):
@@ -55,3 +55,62 @@ class InvalidResetMode(RuntimeError):
         msg += "- 'sync'  - synchronous\n"
         msg += "but was: " + reset
         super().__init__(error(msg))
+
+
+class ModuleDataBitsExceeded(Exception):
+    """! @brief Raised when the specified module data bits are exceeded.
+
+    """
+    def __init__(self, register, reglength, mod_data_length):
+        msg = "Register length exceeded module data length by "
+        msg += str(reglength - mod_data_length)
+        msg += " in register " + register + "\n"
+        msg += 'Module length: ' + str(mod_data_length) + '\n'
+        msg += 'Register length: ' + str(reglength)
+
+        super().__init__(msg)
+
+
+class UndefinedRegisterType(RuntimeError):
+    """! @brief Raised when trying to parse a register type that is not supported.
+
+    """
+    def __init__(self, regtype):
+        msg = "Could not parse register type: " + regtype
+        super().__init__(msg)
+
+
+class UndefinedFieldType(RuntimeError):
+    """! @brief Raised when trying to parse an field type that is not supported.
+
+    """
+    def __init__(self, sig_type):
+        msg = "Could not parse field type: " + sig_type
+        super().__init__(msg)
+
+
+class InvalidRegisterFormat(RuntimeError):
+    """! @brief Raised when register has some unspecified format error.
+
+    """
+    def __init__(self, msg):
+        msg = 'Invalid register format: ' + msg
+        super().__init__(msg)
+
+
+class InvalidFieldFormat(RuntimeError):
+    """! @brief Raised when field format is wrong.
+
+    """
+    def __init__(self, msg):
+        msg = 'Invalid field format: ' + msg
+        super().__init__(msg)
+
+
+class InvalidStallValue(RuntimeError):
+    """! @brief Raised when stall value is wrong.
+
+    """
+    def __init__(self, val):
+        msg = 'Invalid stall value: {}. Must be between 2 and 255.'.format(val)
+        super().__init__(msg)
